create TariffServiceImpl

+++ src/main/java/uzumtech/j_delivery/repository/TariffRepository.java
@@ -3,5 +3,5 @@ package uzumtech.j_delivery.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uzumtech.j_delivery.entity.TariffEntity;
 
-public interface TariffRepository extends JpaRepository<TariffEntity, Integer> {
+public interface TariffRepository extends JpaRepository<TariffEntity, Long> {
 }
